refactor: rename block implementations directory
